feat: Make Calendar::property_value() publicly accessible
There does not seem to be a good reason for making the
Calendar::property_value() method private: the accessed properties array
is public anyway and the method is nothing more than a helper. What's
more, calendar components have exactly a method of this name (fulfilling
a similar purpose) publicly available.
Declare it as public to make it accessible to users, instead of forcing
them to copy and paste it.